refactor(configloader): add all classes of config loader in one file
