+++ connectors-sdk/connectors_sdk/models/configs/__init__.py
@@ -1,14 +1,11 @@
-from connectors_sdk.models.configs.base import BaseConfigModel, _SettingLoader
-from connectors_sdk.models.configs.connector import (
+from connectors_sdk.models.configs.base import (
+    BaseConfigModel,
     BaseConnectorConfig,
     BaseConnectorSettings,
 )
-from connectors_sdk.models.configs.opencti import _OpenCTIConfig
 
 __all__ = [
     BaseConfigModel,
-    _SettingLoader,
-    _OpenCTIConfig,
     BaseConnectorConfig,
     BaseConnectorSettings,
 ]
