+++ connectors-sdk/connectors_sdk/models/configs/base.py
@@ -1,12 +1,24 @@
+import os
 from abc import ABC
+from datetime import timedelta
 from pathlib import Path
+from typing import Literal
 
-from pydantic import BaseModel, ConfigDict
+import __main__
+from connectors_sdk.core.pydantic import ListFromString
+from connectors_sdk.exceptions import (  # ConfigNotFoundError,
+    ConfigError,
+    ConfigValidationError,
+)
+from pydantic import BaseModel, ConfigDict, Field, HttpUrl, ValidationError
 from pydantic_settings import (
     BaseSettings,
     PydanticBaseSettingsSource,
+    SettingsConfigDict,
     YamlConfigSettingsSource,
 )
+
+_MAIN_PATH = os.path.dirname(os.path.abspath(__main__.__file__))
 
 
 class BaseConfigModel(BaseModel, ABC):
@@ -19,6 +31,15 @@ class BaseConfigModel(BaseModel, ABC):
 
 
 class _SettingLoader(BaseSettings):
+    # Setup model config and env vars parsing
+    model_config = SettingsConfigDict(
+        enable_decoding=False,
+        env_nested_delimiter="_",
+        env_nested_max_split=1,
+        env_file=f"{_MAIN_PATH}/../.env",
+        yaml_file=f"{_MAIN_PATH}/../config.yml",
+    )
+
     @classmethod
     def settings_customise_sources(
         cls,
@@ -42,6 +63,68 @@ class _SettingLoader(BaseSettings):
         if Path(settings_cls.model_config["env_file"] or "").is_file():  # type: ignore
             return (dotenv_settings,)
         return (env_settings,)
+
+
+class _OpenCTIConfig(BaseConfigModel):
+    """
+    Define config specific to OpenCTI
+    """
+
+    url: HttpUrl = Field(
+        description="The base URL of the OpenCTI instance.",
+    )
+    token: str = Field(
+        description="The API token to connect to OpenCTI.",
+    )
+
+
+class BaseConnectorConfig(BaseConfigModel, ABC):
+    """
+    Define config specific to a connector
+    """
+
+    id: str = Field(
+        description="A UUID v4 to identify the connector in OpenCTI.",
+    )
+    name: str = Field(
+        description="The name of the connector.",
+    )
+    scope: ListFromString = Field(
+        description="The scope of the connector, e.g. 'flashpoint'."
+    )
+    duration_period: timedelta = Field(
+        description="The period of time to await between two runs of the connector."
+    )
+    log_level: Literal["debug", "info", "warn", "warning", "error", "critical"] = Field(
+        description="The minimum level of logs to display."
+    )
+    # todo : type external import or internal enrichment
+
+
+class BaseConnectorSettings(BaseConfigModel, ABC):
+    """
+    Interface for loading global configuration settings
+    """
+
+    opencti: _OpenCTIConfig = Field(
+        default_factory=_OpenCTIConfig,
+        description="OpenCTI configurations.",
+    )
+    connector: BaseConnectorConfig = Field(
+        default_factory=BaseConnectorConfig,
+        description="Connector configurations.",
+    )
+
+    def __init__(self) -> None:
+        """
+        Wrap BaseConnectorSettings initialization to raise custom exception in case of error.
+        """
+        try:
+            super().__init__()
+        except ValidationError as e:
+            raise ConfigValidationError(".", e) from e
+        except Exception as e:
+            raise ConfigError("Invalid OpenCTI configuration.", e) from e
 
     def model_dump_pycti(self) -> dict:
         """
